# images/swagger.yaml
swagger: "2.0"
info:
  title: Images API
  description: Documentation for Images API
  version: "1.0.0"
host: localhost
basePath: /
schemes:
  - http
consumes:
  - multipart/form-data
produces:
  - application/json
  - image/png
  - image/jpeg
paths:
  /images/covers/{id}:
    post:
      summary: Upload a book cover image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the book cover
          required: true
          type: string
        - name: upload
          in: formData
          description: Image file to upload
          required: true
          type: file
      responses:
        "201":
          description: Successfully uploaded
        "400":
          description: Invalid file type
        "500":
          description: Upload failed
    get:
      summary: Get a book cover image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the book cover
          required: true
          type: string
      responses:
        "200":
          description: Image file
        "404":
          description: File not found
    delete:
      summary: Delete a book cover image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the book cover
          required: true
          type: string
      responses:
        "200":
          description: File deleted
        "404":
          description: File not found
        "500":
          description: Unable to delete file
  /images/avatars/{id}:
    post:
      summary: Upload a user avatar image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the user avatar
          required: true
          type: string
        - name: upload
          in: formData
          description: Image file to upload
          required: true
          type: file
      responses:
        "201":
          description: Successfully uploaded
        "400":
          description: Invalid file type
        "500":
          description: Upload failed
    get:
      summary: Get a user avatar image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the user avatar
          required: true
          type: string
      responses:
        "200":
          description: Image file
        "404":
          description: File not found
    delete:
      summary: Delete a user avatar image
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: ID of the user avatar
          required: true
          type: string
      responses:
        "200":
          description: File deleted
        "404":
          description: File not found
        "500":
          description: Unable to delete file